connected to product detail, and changed price type to float

+++ app/src/main/java/app/appworks/school/stylish/data/OrderDetail.kt
@@ -36,7 +36,7 @@ data class Recipient(
 data class OrderProduct(
     val id: Long,
     val name: String,
-    val price: Int,
+    val price: Float,
     val color: Color,
     val size: String,
     val qty: Long
